refactor(core/query-builder): extract EntityRef/alias/isEntityRef into query-builder/alias (#405)
Pure move; re-exported from SelectQueryBuilder.ts.
